refactor: format 함수 utils 파일로 분리
